Add outputType attribute to blocks.
This allows a global variable to be declared with an expression that
does not immediately have a value. If the expression returns None, the
variable is initialised with a type of the expression block's
outputType, and no value.

An exception is still raised if the expression block has no outputType.