Make runtime input block movement sane.
Important for long running input-type blocks that operate by run()
rather than eval(), such as Bind / StateMonitor controls. This change
means that they can be moved around at runtime in a way that seems to
make sense.

If a running block is disconnected from its parent, it will keep running
unless it is connected to a complete / cancelled / errored parent.

If a cancelled / complete block is connected to a ready block, it will
be reset. If the parent block is running, the child block will be run.
If the parent block is paused, the child block will be paused if it is
running, or reset if it is not paused.