Build Javascript files using Rollup rather than Grunt.
- Change require()s to imports.
- Refactor Block definitions for blocks that define scoped variables to
all use a single mixin. This ensures that the variable is created and
is in the correct scope. It also means that the octopus-python
generator works properly and the built generator js does not require
Blockly itself as a dependency.
- Remove Blockly.Variables. B.V.generateUniqueName was never called.
- FieldTextInput.numberValidator moved into it’s own ‘validators’ file
so that it can be used by the generators without needing to import the
whole of Blockly.
- Remove references to Blockly.Realtime.
- Load Javascript dependencies from CDNs where possible rather than
downloading them with Bower.
- Updates for d3.js v3->v4
- Updates for latest lodash (_.pluck -> _.map)
- Jquery UI minimal custom build included.
- build-dist.py will be moved into a separate repository.